+++ src/checklists/context_processors.py
@@ -18,6 +18,7 @@ def boss_context(request):
     """
     Добавляет переменную is_boss во все шаблоны.
     """
-    if request.user.is_authenticated:
+    # if request.user.is_authenticated:
+    if hasattr(request, "user") and request.user.is_authenticated:
         return {"is_boss": request.user.role in BOSS_ROLES or request.user.is_staff}
     return {"is_boss": False}